Create a new fused_moe_lora_fp8_op.py
