+++ vllm/lora/ops/triton_ops/__init__.py
@@ -10,10 +10,7 @@ from vllm.lora.ops.triton_ops.fused_moe_lora_fp8_op import (
 from vllm.lora.ops.triton_ops.fused_moe_lora_op import (
     fused_moe_lora,
     fused_moe_lora_expand,
-    fused_moe_lora_expand_fp8,
-    fused_moe_lora_fp8,
     fused_moe_lora_shrink,
-    fused_moe_lora_shrink_fp8,
 )
 from vllm.lora.ops.triton_ops.lora_expand_op import lora_expand
 from vllm.lora.ops.triton_ops.lora_kernel_metadata import LoRAKernelMeta
